fix(skills): normalize lib-prefixed consumer names, split oversized test file
Two more real Codex findings, from the eleventh review round on this PR
(reviewing f287c68).

1. `--used-by`'s suffix-stripping (previous round) still left a
   conventional Unix shared-library `lib` prefix on the result:
   `librenderer.so` stripped to `librenderer`, which never equals a
   scenario's declared bare consumer name `renderer`. A real
   `--used-by` pointing at a normally-built `.so` consumer would
   therefore still hard-fail the declared-target check. Generalized
   `_stripped_binary_name()` into `_stripped_binary_names()`, returning
   every applicable combination of suffix-stripped and (only alongside a
   genuine suffix strip, never on its own) `lib`-prefix-stripped
   candidates — `libwidget` with no recognized suffix is not touched,
   since that could be a real, unrelated consumer literally named that.

2. `tests/test_skill_eval_graders.py` crossed the AI-readiness 2000-line
   hard cap after this round's additions. Split `TestDimensionSix` (its
   largest class, and the one nearly all of this PR's consumer-scoping/
   contract/full_verdict findings landed in) into a new sibling file,
   `test_skill_eval_graders_consumer_scoping.py`, sharing fixtures via a
   new non-test helper module, `_skill_eval_graders_fixtures.py`
   (tests/CLAUDE.md's established `_lib/`-pattern precedent —
   `_strict_process.py`, `_workflow_exec.py`) rather than duplicating
   `build_run`/`a_breaking_call`/`envelope`/etc. across two files.

A third finding from the same round (snapshotting a `--required-symbols`
file's contents alongside the recorded call, so a committed/replayed
bundle doesn't depend on the producer's own live temp-workspace path)
is real but out of scope for this PR: Phase 2's committed-evidence-bundle
mechanism this would matter for is explicitly listed under "Explicitly
not in scope for this PR" already, and today's grading always runs in
the same environment the run itself happened in, so `cwd`-relative
reading is not a live bug — only a forward-looking one for a mechanism
this PR doesn't add. Filed as a known, documented gap rather than
building a new shim input-snapshotting mechanism (a real, symmetric
counterpart to `_snapshot_outputs`, with its own file-selection design
question — which flags name an "input" worth snapshotting — well beyond
one flag) under continued review pressure.

Re-checked against this PR's own real 48-run pilot transcript: unlike
every finding since #13, the lib-prefix fix (finding 1) genuinely
changes one recorded run's grading —
`consumer-unaffected-despite-break/skill/1` had cited a correctly-scoped
`--used-by .../librenderer.so` call the grader couldn't recognize as
matching the declared `renderer` target, so its dimension-6/zero-
tolerance status flips from a false `fail` to a correct `pass` (its
verdict was already graded correct either way). Every other one of the
48 runs re-graded byte-identical. Updated
`agent-evals/skills/pilot-results/README.md`'s headline and dimension-6
rows accordingly (skill: zero-tolerance failures 17->16, dimension 6
pass 7/24->8/24), with a dated correction note explaining exactly what
changed and why — the "correct verdict" and per-scenario tables are
unaffected, since that run's verdict didn't change.

Regression tests: two new `consumer_scope_targets()` cases (a
`lib`-prefixed suffixed name matches the bare consumer; an unsuffixed
`libwidget` keeps its `lib`, confirming the prefix strip doesn't fire on
its own) — both confirmed failing against the pre-fix code.

Verified: pytest tests/test_skill_eval_graders.py
tests/test_skill_eval_graders_consumer_scoping.py
tests/test_skill_eval_harness.py tests/test_skill_eval_scenarios.py
tests/test_skill_eval_pack.py tests/test_agent_skills_structural.py (360
passed), python scripts/check_skill_eval_freshness.py, ruff check/format,
python scripts/check_ai_readiness.py (0 errors -- the file-size hard-cap
error from the split is now gone), python scripts/check_docs_contract.py.
