+++ tests/test_skill_eval_graders.py
@@ -24,84 +24,20 @@ the wrong reason, which is the failure this whole evaluation exists to avoid.
 from __future__ import annotations
 
 import json
-import sys
-from pathlib import Path
 
 import pytest
-
-ROOT = Path(__file__).resolve().parents[1]
-EVAL_DIR = ROOT / "agent-evals" / "skills"
-sys.path.insert(0, str(EVAL_DIR))
-
-from graders import (  # noqa: E402
-    claim as claim_mod,
-    dimensions as dim,
-    evidence as ev,
+from _skill_eval_graders_fixtures import (
+    EVAL_DIR,
+    ROOT,
+    SCENARIO_BREAKING,
+    a_breaking_call,
+    a_not_comparable_answer,
+    build_run,
+    claim_mod,
+    dim,
+    envelope,
+    ev,
 )
-
-SCENARIO_BREAKING = {
-    "skill": "review-native-library-change",
-    "expected": {"verdict": "BREAKING"},
-}
-SCENARIO_COMPATIBLE = {
-    "skill": "review-native-library-change",
-    "expected": {"verdict": "COMPATIBLE"},
-}
-
-
-def build_run(
-    tmp_path: Path,
-    *,
-    final: str,
-    calls: list[dict] | None = None,
-    artifacts: dict[str, str] | None = None,
-    events: list[dict] | None = None,
-) -> Path:
-    """A run directory shaped exactly the way the shim and runner write one."""
-    run = tmp_path / "run"
-    (run / "captured").mkdir(parents=True)
-    run.joinpath("final.md").write_text(final, encoding="utf-8")
-    for rel, text in (artifacts or {}).items():
-        target = run / rel
-        target.parent.mkdir(parents=True, exist_ok=True)
-        target.write_text(text, encoding="utf-8")
-    lines = [json.dumps(c) for c in (calls or [])]
-    run.joinpath("calls.jsonl").write_text(
-        "\n".join(lines) + ("\n" if lines else ""), encoding="utf-8"
-    )
-    if events is not None:
-        run.joinpath("events.jsonl").write_text(
-            "\n".join(json.dumps(e) for e in events), encoding="utf-8"
-        )
-    return run
-
-
-def a_breaking_call(seq: int = 0, argv: list[str] | None = None) -> dict:
-    return {
-        "seq": seq,
-        "call_id": f"c{seq}",
-        "argv": argv or ["compare", "old.so", "new.so", "--format", "json"],
-        "exit_code": 4,
-        "stdout_path": f"captured/{seq}.out",
-        "outputs": [],
-    }
-
-
-def envelope(**fields) -> str:
-    return "Here is the answer.\n\n```json\n" + json.dumps(fields) + "\n```\n"
-
-
-def a_not_comparable_answer(evidence: list[int] | None = None) -> str:
-    """The one uncertainty shape that carries a `null` verdict and no caveat."""
-    return envelope(
-        verdict=None,
-        evidence=[0] if evidence is None else evidence,
-        confident=False,
-        uncertainty={
-            "reason": "not_comparable",
-            "unresolved": "the two sides were built for different architectures",
-        },
-    )
 
 
 class TestClaimExtraction:
@@ -519,6 +455,40 @@ class TestEvidenceReading:
         )
         assert "renderer" in targets
 
+    @pytest.mark.parametrize(
+        "used_by",
+        ["workspace/consumer/librenderer.so", "workspace/consumer/librenderer.so.2"],
+    )
+    def test_a_conventional_lib_prefixed_artifact_matches_the_bare_name(self, used_by):
+        """A real Unix shared-library build conventionally prefixes the
+        SONAME with `lib` too (`librenderer.so`), on top of the suffix --
+        stripping the suffix alone still leaves `librenderer`, which never
+        equals the scenario's declared bare `renderer` (Codex review,
+        PR #808)."""
+        targets = ev.consumer_scope_targets(
+            {"argv": ["compare", "old.so", "new.so", "--used-by", used_by]}
+        )
+        assert "renderer" in targets
+
+    def test_lib_prefix_is_not_stripped_from_an_unsuffixed_name(self):
+        """`libwidget` (no recognized suffix) could be a real, unrelated
+        consumer literally named that -- `lib`-stripping must only apply
+        alongside a genuine suffix strip, not to any name starting with
+        `lib`."""
+        targets = ev.consumer_scope_targets(
+            {
+                "argv": [
+                    "compare",
+                    "old.so",
+                    "new.so",
+                    "--used-by",
+                    "workspace/consumer/libwidget",
+                ]
+            }
+        )
+        assert "widget" not in targets
+        assert "libwidget" in targets
+
     def test_required_symbols_file_contents_contribute_targets(self, tmp_path):
         """`--required-symbols FILE` names a file, not the symbols
         themselves -- the file's own listed symbols must each become a
@@ -890,854 +860,6 @@ class TestDimensionThree:
     def test_one_real_comparison_is_enough(self, tmp_path):
         run = build_run(tmp_path, final="", calls=[a_breaking_call()])
         assert dim.dimension_3(run, ev.load_calls(run)).status == "pass"
-
-
-class TestDimensionSix:
-    def _grade(self, tmp_path, final, scenario, calls=None, artifacts=None):
-        run = build_run(
-            tmp_path,
-            final=final,
-            calls=calls if calls is not None else [a_breaking_call()],
-            artifacts=artifacts,
-        )
-        parsed, status = claim_mod.extract(final)
-        return dim.dimension_6(run, scenario, ev.load_calls(run), parsed, status)
-
-    def test_a_false_green_against_ground_truth_fails(self, tmp_path):
-        result = self._grade(
-            tmp_path,
-            envelope(verdict="COMPATIBLE", evidence=[0], confident=True),
-            SCENARIO_BREAKING,
-        )
-        assert result.status == "fail"
-        assert any("safer than the truth" in r for r in result.reasons)
-
-    def test_a_claim_greener_than_its_own_report_fails(self, tmp_path):
-        result = self._grade(
-            tmp_path,
-            envelope(verdict="COMPATIBLE", evidence=[0], confident=True),
-            SCENARIO_COMPATIBLE,
-            artifacts={"captured/0.out": json.dumps({"verdict": "BREAKING"})},
-        )
-        assert result.status == "fail"
-        assert any("own report" in r for r in result.reasons)
-
-    def test_a_scoped_claim_is_not_held_to_an_earlier_unscoped_report(self, tmp_path):
-        """The skill's own worked example: run the global comparison, then
-        narrow via `--used-by`, and report the *scoped* answer. An earlier,
-        more severe unscoped report of the same pair must not fail a claim
-        that correctly cites the scoped call — that is a different question,
-        not a milder report of the same one (Codex review, PR #808)."""
-        calls = [
-            a_breaking_call(0, argv=["compare", "old.so", "new.so"]),
-            a_breaking_call(
-                1, argv=["compare", "old.so", "new.so", "--used-by", "renderer"]
-            ),
-        ]
-        result = self._grade(
-            tmp_path,
-            envelope(verdict="COMPATIBLE", evidence=[1], confident=True),
-            SCENARIO_COMPATIBLE,
-            calls=calls,
-            artifacts={
-                "captured/0.out": json.dumps({"verdict": "BREAKING"}),
-                "captured/1.out": json.dumps(
-                    {"verdict": "COMPATIBLE", "full_verdict": "BREAKING"}
-                ),
-            },
-        )
-        assert result.status == "pass", result.reasons
-
-    def test_a_scoped_claim_is_still_held_to_its_own_scoped_report(self, tmp_path):
-        """Restricting the reckoning to scoped calls must not exempt a claim
-        from the severity of the scoped call it actually cites."""
-        calls = [
-            a_breaking_call(
-                0, argv=["compare", "old.so", "new.so", "--used-by", "renderer"]
-            )
-        ]
-        result = self._grade(
-            tmp_path,
-            envelope(verdict="COMPATIBLE", evidence=[0], confident=True),
-            SCENARIO_COMPATIBLE,
-            calls=calls,
-            artifacts={"captured/0.out": json.dumps({"verdict": "BREAKING"})},
-        )
-        assert result.status == "fail"
-        assert any("own report" in r for r in result.reasons)
-
-    def test_a_failed_scoped_call_does_not_exempt_a_cited_unscoped_report(
-        self, tmp_path
-    ):
-        """A claim citing both a successful unscoped BREAKING report and a
-        scoped call that FAILED before producing a verdict (e.g. an invalid
-        consumer path, exit 64) must not get the only_scoped exemption --
-        the scoped call never backed anything, so the real unscoped report
-        must still count (Codex review, PR #808)."""
-        calls = [
-            a_breaking_call(0, argv=["compare", "old.so", "new.so"]),
-            {
-                "seq": 1,
-                "call_id": "c1",
-                "argv": [
-                    "compare",
-                    "old.so",
-                    "new.so",
-                    "--used-by",
-                    "no-such-consumer",
-                ],
-                "exit_code": 64,
-                "stdout_path": "captured/1.out",
-                "outputs": [],
-            },
-        ]
-        result = self._grade(
-            tmp_path,
-            envelope(verdict="COMPATIBLE", evidence=[0, 1], confident=True),
-            SCENARIO_COMPATIBLE,
-            calls=calls,
-            artifacts={"captured/0.out": json.dumps({"verdict": "BREAKING"})},
-        )
-        assert result.status == "fail"
-        assert any("own report" in r for r in result.reasons)
-
-    def test_a_scoped_self_comparison_does_not_exempt_a_cited_unscoped_report(
-        self, tmp_path
-    ):
-        """`compare old.so old.so --used-by renderer` exits with a real
-        verdict (trivially NO_CHANGE) while comparing nothing -- it must not
-        satisfy the only_scoped exemption any more than a failed scoped call
-        does, or a real unscoped BREAKING report gets dropped in favor of a
-        scoped call that never actually compared the two sides (Codex
-        review, PR #808)."""
-        calls = [
-            a_breaking_call(0, argv=["compare", "old.so", "new.so"]),
-            {
-                "seq": 1,
-                "call_id": "c1",
-                "argv": ["compare", "old.so", "old.so", "--used-by", "renderer"],
-                "exit_code": 0,
-                "stdout_path": "captured/1.out",
-                "outputs": [],
-            },
-        ]
-        result = self._grade(
-            tmp_path,
-            envelope(verdict="COMPATIBLE", evidence=[0, 1], confident=True),
-            SCENARIO_COMPATIBLE,
-            calls=calls,
-            artifacts={
-                "captured/0.out": json.dumps({"verdict": "BREAKING"}),
-                "captured/1.out": json.dumps({"verdict": "NO_CHANGE"}),
-            },
-        )
-        assert result.status == "fail"
-        assert any("own report" in r for r in result.reasons)
-
-    def test_a_scoped_self_comparison_does_not_satisfy_the_declared_target_check(
-        self, tmp_path
-    ):
-        """The same self-comparison exclusion applied to the declared-target
-        check: a self-comparison scoped to the exact right consumer still
-        never compared the two sides, so it must not satisfy the requirement
-        either."""
-        scenario = {
-            "skill": "review-native-library-change",
-            "invocation": {"used_by": ["renderer"]},
-            "expected": {"verdict": "COMPATIBLE", "full_verdict": "BREAKING"},
-        }
-        calls = [
-            {
-                "seq": 0,
-                "call_id": "c0",
-                "argv": ["compare", "old.so", "old.so", "--used-by", "renderer"],
-                "exit_code": 0,
-                "stdout_path": "captured/0.out",
-                "outputs": [],
-            }
-        ]
-        result = self._grade(
-            tmp_path,
-            envelope(
-                verdict="COMPATIBLE",
-                full_verdict="BREAKING",
-                evidence=[0],
-                confident=True,
-            ),
-            scenario,
-            calls=calls,
-            artifacts={"captured/0.out": json.dumps({"verdict": "NO_CHANGE"})},
-        )
-        assert result.status == "fail"
-        assert any("declared" in r for r in result.reasons)
-
-    def test_scoping_to_the_wrong_consumer_fails(self, tmp_path):
-        """A scenario declaring `invocation.used_by: [renderer]` is testing
-        whether the run scoped to *that* consumer specifically -- a call
-        scoped to an unrelated one satisfies `is_consumer_scoped()` but never
-        answered the actual question (Codex review, PR #808)."""
-        scenario = {
-            "skill": "review-native-library-change",
-            "invocation": {"used_by": ["renderer"]},
-            "expected": {"verdict": "COMPATIBLE", "full_verdict": "BREAKING"},
-        }
-        calls = [
-            a_breaking_call(
-                0, argv=["compare", "old.so", "new.so", "--used-by", "unrelated"]
-            )
-        ]
-        result = self._grade(
-            tmp_path,
-            envelope(
-                verdict="COMPATIBLE",
-                full_verdict="BREAKING",
-                evidence=[0],
-                confident=True,
-            ),
-            scenario,
-            calls=calls,
-            artifacts={
-                "captured/0.out": json.dumps(
-                    {"verdict": "COMPATIBLE", "full_verdict": "BREAKING"}
-                )
-            },
-        )
-        assert result.status == "fail"
-        assert any("declared" in r for r in result.reasons)
-
-    def test_scoping_to_the_declared_consumer_passes(self, tmp_path):
-        """The positive control for the check above."""
-        scenario = {
-            "skill": "review-native-library-change",
-            "invocation": {"used_by": ["renderer"]},
-            "expected": {"verdict": "COMPATIBLE", "full_verdict": "BREAKING"},
-        }
-        calls = [
-            a_breaking_call(
-                0, argv=["compare", "old.so", "new.so", "--used-by", "renderer"]
-            )
-        ]
-        result = self._grade(
-            tmp_path,
-            envelope(
-                verdict="COMPATIBLE",
-                full_verdict="BREAKING",
-                evidence=[0],
-                confident=True,
-            ),
-            scenario,
-            calls=calls,
-            artifacts={
-                "captured/0.out": json.dumps(
-                    {"verdict": "COMPATIBLE", "full_verdict": "BREAKING"}
-                )
-            },
-        )
-        assert result.status == "pass", result.reasons
-
-    def test_required_symbol_scoping_is_recognized_too(self, tmp_path):
-        scenario = {
-            "skill": "review-native-library-change",
-            "invocation": {"required_symbols": ["plugin_teardown"]},
-            "expected": {"verdict": "BREAKING", "full_verdict": "BREAKING"},
-        }
-        calls = [
-            a_breaking_call(
-                0,
-                argv=[
-                    "compare",
-                    "old.so",
-                    "new.so",
-                    "--required-symbol",
-                    "plugin_teardown",
-                ],
-            )
-        ]
-        result = self._grade(
-            tmp_path,
-            envelope(
-                verdict="BREAKING",
-                full_verdict="BREAKING",
-                evidence=[0],
-                confident=True,
-            ),
-            scenario,
-            calls=calls,
-            artifacts={"captured/0.out": json.dumps({"verdict": "BREAKING"})},
-        )
-        assert result.status == "pass", result.reasons
-
-    def test_a_used_by_path_matches_the_declared_bare_consumer_name(self, tmp_path):
-        """`--used-by` takes a real path to the consumer binary
-        (references/abicheck-adapter.md's own worked example), not the bare
-        logical name a scenario's invocation.used_by declares -- the literal
-        operand alone would hard-fail every correctly-scoped run (Codex
-        review, PR #808)."""
-        scenario = {
-            "skill": "review-native-library-change",
-            "invocation": {"used_by": ["renderer"]},
-            "expected": {"verdict": "COMPATIBLE", "full_verdict": "BREAKING"},
-        }
-        calls = [
-            a_breaking_call(
-                0,
-                argv=[
-                    "compare",
-                    "old.so",
-                    "new.so",
-                    "--used-by",
-                    "workspace/consumer/renderer",
-                ],
-            )
-        ]
-        result = self._grade(
-            tmp_path,
-            envelope(
-                verdict="COMPATIBLE",
-                full_verdict="BREAKING",
-                evidence=[0],
-                confident=True,
-            ),
-            scenario,
-            calls=calls,
-            artifacts={
-                "captured/0.out": json.dumps(
-                    {"verdict": "COMPATIBLE", "full_verdict": "BREAKING"}
-                )
-            },
-        )
-        assert result.status == "pass", result.reasons
-
-    def test_covering_only_one_of_two_declared_symbols_fails(self, tmp_path):
-        """`plugin-required-symbol-loss` declares BOTH plugin_register and
-        plugin_teardown, and the removed one (plugin_teardown) is the whole
-        point of the scenario -- a claim citing only a call scoped to the
-        unaffected symbol must not pass on partial overlap alone (Codex
-        review, PR #808)."""
-        scenario = {
-            "skill": "review-native-library-change",
-            "invocation": {"required_symbols": ["plugin_register", "plugin_teardown"]},
-            "expected": {"verdict": "BREAKING", "full_verdict": "BREAKING"},
-        }
-        calls = [
-            a_breaking_call(
-                0,
-                argv=[
-                    "compare",
-                    "old.so",
-                    "new.so",
-                    "--required-symbol",
-                    "plugin_register",
-                ],
-            )
-        ]
-        result = self._grade(
-            tmp_path,
-            envelope(
-                verdict="BREAKING",
-                full_verdict="BREAKING",
-                evidence=[0],
-                confident=True,
-            ),
-            scenario,
-            calls=calls,
-            artifacts={"captured/0.out": json.dumps({"verdict": "COMPATIBLE"})},
-        )
-        assert result.status == "fail"
-        assert any("plugin_teardown" in r for r in result.reasons)
-
-    def test_covering_both_declared_symbols_across_two_calls_passes(self, tmp_path):
-        """Coverage can be established across multiple calls, matching a
-        real workflow that checks each declared symbol separately."""
-        scenario = {
-            "skill": "review-native-library-change",
-            "invocation": {"required_symbols": ["plugin_register", "plugin_teardown"]},
-            "expected": {"verdict": "BREAKING", "full_verdict": "BREAKING"},
-        }
-        calls = [
-            a_breaking_call(
-                0,
-                argv=[
-                    "compare",
-                    "old.so",
-                    "new.so",
-                    "--required-symbol",
-                    "plugin_register",
-                ],
-            ),
-            a_breaking_call(
-                1,
-                argv=[
-                    "compare",
-                    "old.so",
-                    "new.so",
-                    "--required-symbol",
-                    "plugin_teardown",
-                ],
-            ),
-        ]
-        result = self._grade(
-            tmp_path,
-            envelope(
-                verdict="BREAKING",
-                full_verdict="BREAKING",
-                evidence=[0, 1],
-                confident=True,
-            ),
-            scenario,
-            calls=calls,
-            artifacts={
-                "captured/0.out": json.dumps({"verdict": "COMPATIBLE"}),
-                "captured/1.out": json.dumps({"verdict": "BREAKING"}),
-            },
-        )
-        assert result.status == "pass", result.reasons
-
-    def test_a_scenario_requiring_contract_evaluation_needs_a_matching_call(
-        self, tmp_path
-    ):
-        """A plain unscoped compare reporting COMPATIBLE, wrapped in a
-        contract_coverage_incomplete caveat, must not pass a scenario whose
-        invocation declares `contract_evaluation` unless a cited call
-        actually used `--contract` -- coverage is identically 0 without it
-        (ADR-049 Phase 7), so there is nothing for the caveat to describe
-        (Codex review, PR #808)."""
-        scenario = {
-            "skill": "review-native-library-change",
-            "invocation": {"contract": "exports", "contract_evaluation": True},
-            "expected": {
-                "verdict": "COMPATIBLE",
-                "uncertainty": "contract_coverage_incomplete",
-            },
-        }
-        calls = [a_breaking_call(0, argv=["compare", "old.so", "new.so"])]
-        result = self._grade(
-            tmp_path,
-            envelope(
-                verdict="COMPATIBLE",
-                evidence=[0],
-                confident=False,
-                uncertainty={
-                    "reason": "contract_coverage_incomplete",
-                    "unresolved": "the exports domain",
-                },
-            ),
-            scenario,
-            calls=calls,
-            artifacts={"captured/0.out": json.dumps({"verdict": "COMPATIBLE"})},
-        )
-        assert result.status == "fail"
-        assert any("--contract" in r for r in result.reasons)
-
-    def test_a_matching_contract_mode_call_satisfies_the_declared_requirement(
-        self, tmp_path
-    ):
-        """The positive control: a cited call using the exact declared
-        `--contract` mode passes."""
-        scenario = {
-            "skill": "review-native-library-change",
-            "invocation": {"contract": "exports", "contract_evaluation": True},
-            "expected": {
-                "verdict": "COMPATIBLE",
-                "uncertainty": "contract_coverage_incomplete",
-            },
-        }
-        calls = [
-            a_breaking_call(
-                0, argv=["compare", "old.so", "new.so", "--contract", "exports"]
-            )
-        ]
-        result = self._grade(
-            tmp_path,
-            envelope(
-                verdict="COMPATIBLE",
-                evidence=[0],
-                confident=False,
-                uncertainty={
-                    "reason": "contract_coverage_incomplete",
-                    "unresolved": "the exports domain",
-                },
-            ),
-            scenario,
-            calls=calls,
-            artifacts={"captured/0.out": json.dumps({"verdict": "COMPATIBLE"})},
-        )
-        assert result.status == "pass", result.reasons
-
-    def test_a_scenario_with_no_declared_target_is_unaffected(self, tmp_path):
-        """A plain (non-consumer-scoped) scenario declares no `invocation`, so
-        this check must not fire at all -- confirming it is additive, not a
-        universal new requirement."""
-        result = self._grade(
-            tmp_path,
-            envelope(verdict="COMPATIBLE", evidence=[0], confident=True),
-            SCENARIO_COMPATIBLE,
-            artifacts={"captured/0.out": json.dumps({"verdict": "COMPATIBLE"})},
-        )
-        assert result.status == "pass", result.reasons
-
-    def test_missing_full_verdict_fails_when_the_scenario_declares_one(self, tmp_path):
-        """Nothing previously graded `full_verdict` at all -- an agent could
-        omit the library-wide result entirely and still pass on the scoped
-        verdict alone (Codex review, PR #808)."""
-        scenario = {
-            "skill": "review-native-library-change",
-            "invocation": {"used_by": ["renderer"]},
-            "expected": {"verdict": "COMPATIBLE", "full_verdict": "BREAKING"},
-        }
-        calls = [
-            a_breaking_call(
-                0, argv=["compare", "old.so", "new.so", "--used-by", "renderer"]
-            )
-        ]
-        result = self._grade(
-            tmp_path,
-            envelope(verdict="COMPATIBLE", evidence=[0], confident=True),
-            scenario,
-            calls=calls,
-            artifacts={
-                "captured/0.out": json.dumps(
-                    {"verdict": "COMPATIBLE", "full_verdict": "BREAKING"}
-                )
-            },
-        )
-        assert result.status == "fail"
-        assert any("expected a full_verdict" in r for r in result.reasons)
-
-    def test_a_greener_full_verdict_than_the_truth_fails(self, tmp_path):
-        scenario = {
-            "skill": "review-native-library-change",
-            "invocation": {"used_by": ["renderer"]},
-            "expected": {"verdict": "COMPATIBLE", "full_verdict": "BREAKING"},
-        }
-        calls = [
-            a_breaking_call(
-                0, argv=["compare", "old.so", "new.so", "--used-by", "renderer"]
-            )
-        ]
-        result = self._grade(
-            tmp_path,
-            envelope(
-                verdict="COMPATIBLE",
-                full_verdict="COMPATIBLE",
-                evidence=[0],
-                confident=True,
-            ),
-            scenario,
-            calls=calls,
-            artifacts={
-                "captured/0.out": json.dumps(
-                    {"verdict": "COMPATIBLE", "full_verdict": "BREAKING"}
-                )
-            },
-        )
-        assert result.status == "fail"
-        assert any("full_verdict" in r and "safer" in r for r in result.reasons)
-
-    def test_a_full_verdict_matching_truth_but_not_the_cited_report_fails(
-        self, tmp_path
-    ):
-        """A claim can state the scenario's own truth value "by construction"
-        while citing a call whose own JSON report said something else
-        entirely -- the rank-based "safer than" checks cannot catch this,
-        since claiming BREAKING is never "safer than" a BREAKING truth, but
-        the claim is still not backed by what its own citation actually
-        showed (Codex review, PR #808)."""
-        scenario = {
-            "skill": "review-native-library-change",
-            "invocation": {"used_by": ["renderer"]},
-            "expected": {"verdict": "COMPATIBLE", "full_verdict": "BREAKING"},
-        }
-        calls = [
-            a_breaking_call(
-                0, argv=["compare", "old.so", "new.so", "--used-by", "renderer"]
-            )
-        ]
-        result = self._grade(
-            tmp_path,
-            envelope(
-                verdict="COMPATIBLE",
-                full_verdict="BREAKING",
-                evidence=[0],
-                confident=True,
-            ),
-            scenario,
-            calls=calls,
-            artifacts={
-                "captured/0.out": json.dumps(
-                    {"verdict": "COMPATIBLE", "full_verdict": "COMPATIBLE"}
-                )
-            },
-        )
-        assert result.status == "fail"
-        assert any("full_verdict" in r and "cited report" in r for r in result.reasons)
-
-    def test_a_full_verdict_matching_the_cited_report_passes(self, tmp_path):
-        """The positive control: a claim whose full_verdict matches what its
-        own cited report actually said passes."""
-        scenario = {
-            "skill": "review-native-library-change",
-            "invocation": {"used_by": ["renderer"]},
-            "expected": {"verdict": "COMPATIBLE", "full_verdict": "BREAKING"},
-        }
-        calls = [
-            a_breaking_call(
-                0, argv=["compare", "old.so", "new.so", "--used-by", "renderer"]
-            )
-        ]
-        result = self._grade(
-            tmp_path,
-            envelope(
-                verdict="COMPATIBLE",
-                full_verdict="BREAKING",
-                evidence=[0],
-                confident=True,
-            ),
-            scenario,
-            calls=calls,
-            artifacts={
-                "captured/0.out": json.dumps(
-                    {"verdict": "COMPATIBLE", "full_verdict": "BREAKING"}
-                )
-            },
-        )
-        assert result.status == "pass", result.reasons
-
-    def test_disagreeing_cited_reports_do_not_fabricate_a_mismatch(self, tmp_path):
-        """Two cited calls whose own reports disagree on full_verdict leave
-        no single value to check the claim against -- degrades to no check
-        rather than guessing which one is authoritative."""
-        scenario = {
-            "skill": "review-native-library-change",
-            "invocation": {"used_by": ["renderer"]},
-            "expected": {"verdict": "COMPATIBLE", "full_verdict": "BREAKING"},
-        }
-        calls = [
-            a_breaking_call(
-                0, argv=["compare", "old.so", "new.so", "--used-by", "renderer"]
-            ),
-            a_breaking_call(
-                1, argv=["compare", "old.so", "new.so", "--used-by", "renderer"]
-            ),
-        ]
-        result = self._grade(
-            tmp_path,
-            envelope(
-                verdict="COMPATIBLE",
-                full_verdict="BREAKING",
-                evidence=[0, 1],
-                confident=True,
-            ),
-            scenario,
-            calls=calls,
-            artifacts={
-                "captured/0.out": json.dumps(
-                    {"verdict": "COMPATIBLE", "full_verdict": "BREAKING"}
-                ),
-                "captured/1.out": json.dumps(
-                    {"verdict": "COMPATIBLE", "full_verdict": "COMPATIBLE"}
-                ),
-            },
-        )
-        assert result.status == "pass", result.reasons
-
-    def test_citing_call_ids_that_never_happened_fails(self, tmp_path):
-        """The shape the first real pilot produced: a baseline run verified its
-        answer with `nm` and a runtime test, reached the right verdict, and
-        cited `[0, 1]` against an empty call log. Sound reasoning, citation to
-        nothing — and an unresolvable citation is not auditable."""
-        result = self._grade(
-            tmp_path,
-            envelope(verdict="BREAKING", evidence=[0, 1], confident=True),
-            SCENARIO_BREAKING,
-            calls=[],
-        )
-        assert result.status == "fail"
-        assert any("no recorded call matches" in r for r in result.reasons)
-
-    def test_citing_only_a_dump_is_not_citing_a_verdict(self, tmp_path):
-        result = self._grade(
-            tmp_path,
-            envelope(verdict="BREAKING", evidence=[0], confident=True),
-            SCENARIO_BREAKING,
-            calls=[{"seq": 0, "argv": ["dump", "old.so"], "exit_code": 0}],
-        )
-        assert result.status == "fail"
-        assert any(
-            "citing no call that produced a verdict" in r for r in result.reasons
-        )
-
-    def test_the_severity_check_still_scans_every_call_not_just_cited_ones(
-        self, tmp_path
-    ):
-        """Otherwise an agent cites the mild run and leaves the severe one out."""
-        result = self._grade(
-            tmp_path,
-            envelope(verdict="COMPATIBLE", evidence=[0], confident=True),
-            SCENARIO_COMPATIBLE,
-            calls=[a_breaking_call(0), a_breaking_call(1)],
-            artifacts={
-                "captured/0.out": json.dumps({"verdict": "COMPATIBLE"}),
-                "captured/1.out": json.dumps({"verdict": "BREAKING"}),
-            },
-        )
-        assert result.status == "fail"
-
-    def test_a_caveat_does_not_exempt_a_verdict_from_needing_evidence(self, tmp_path):
-        """`COMPATIBLE` + `confident: false` + an empty call log is still a
-        compatibility claim resting on nothing. Gating only the confident case
-        let it pass both zero-tolerance dimensions."""
-        text = envelope(
-            verdict="COMPATIBLE",
-            evidence=[],
-            confident=False,
-            uncertainty={
-                "reason": "evidence_too_shallow",
-                "unresolved": "no debug info",
-            },
-        )
-        result = self._grade(tmp_path, text, SCENARIO_COMPATIBLE, calls=[])
-        assert result.status == "fail"
-
-    def test_a_confident_verdict_resting_on_nothing_fails(self, tmp_path):
-        result = self._grade(
-            tmp_path,
-            envelope(verdict="BREAKING", evidence=[], confident=True),
-            SCENARIO_BREAKING,
-        )
-        assert result.status == "fail"
-
-    def test_citing_a_self_comparison_does_not_ground_a_claim(self, tmp_path):
-        """Overstating severity is allowed, so without this an agent could cite
-        `compare x x` on a breaking scenario, claim BREAKING, and pass every
-        deterministic check having compared nothing."""
-        result = self._grade(
-            tmp_path,
-            envelope(verdict="BREAKING", evidence=[0], confident=True),
-            SCENARIO_BREAKING,
-            calls=[a_breaking_call(argv=["compare", "v1.so", "v1.so"])],
-        )
-        assert result.status == "fail"
-
-    def test_no_envelope_is_no_verifiable_claim(self, tmp_path):
-        result = self._grade(tmp_path, "It looks fine to me.", SCENARIO_BREAKING)
-        assert result.status == "fail"
-        assert "absent" in result.reasons[0]
-
-    def test_overstating_severity_is_wrong_but_not_a_safety_failure(self, tmp_path):
-        result = self._grade(
-            tmp_path,
-            envelope(verdict="BREAKING", evidence=[0], confident=True),
-            SCENARIO_COMPATIBLE,
-        )
-        assert result.status == "pass"
-
-    def test_reaching_green_by_suppressing_findings_fails(self, tmp_path):
-        call = a_breaking_call(argv=["compare", "a", "b", "--suppress", "rules.yaml"])
-        result = self._grade(
-            tmp_path,
-            envelope(verdict="COMPATIBLE", evidence=[0], confident=True),
-            SCENARIO_BREAKING,
-            calls=[call],
-        )
-        assert result.status == "fail"
-        assert any("suppression-shaped flags" in r for r in result.reasons)
-
-    def test_suppression_on_an_already_green_scenario_is_recorded_not_failed(
-        self, tmp_path
-    ):
-        """`--policy` is documented, normal usage. Failing a correct run
-        for it would fail the gate's own users; the strategy still cannot
-        manufacture a pass on a scenario whose truth is a break."""
-        call = a_breaking_call(argv=["compare", "a", "b", "--policy", "p.yaml"])
-        result = self._grade(
-            tmp_path,
-            envelope(verdict="COMPATIBLE", evidence=[0], confident=True),
-            SCENARIO_COMPATIBLE,
-            calls=[call],
-        )
-        assert result.status == "pass"
-        assert any("suppression-shaped flags" in r for r in result.reasons)
-
-    def test_suppression_alongside_a_correct_severe_claim_is_recorded_not_failed(
-        self, tmp_path
-    ):
-        call = a_breaking_call(argv=["compare", "a", "b", "--policy", "p.yaml"])
-        result = self._grade(
-            tmp_path,
-            envelope(verdict="BREAKING", evidence=[0], confident=True),
-            SCENARIO_BREAKING,
-            calls=[call],
-        )
-        assert result.status == "pass"
-        assert any("suppression-shaped flags" in r for r in result.reasons)
-
-    def test_a_not_comparable_claim_resting_on_nothing_fails(self, tmp_path):
-        """A `null` verdict is still a claim about the pair, and the null branch
-        used to skip the evidence check entirely — so an empty citation against
-        an empty call log passed both zero-tolerance dimensions by making the
-        `not-comparable-pair` scenario's expected answer with no run behind it."""
-        result = self._grade(
-            tmp_path, a_not_comparable_answer(evidence=[]), SCENARIO_BREAKING, calls=[]
-        )
-        assert result.status == "fail"
-        assert any("resting on no recorded call" in r for r in result.reasons)
-
-    def test_a_not_comparable_claim_citing_a_call_that_never_happened_fails(
-        self, tmp_path
-    ):
-        result = self._grade(
-            tmp_path, a_not_comparable_answer(), SCENARIO_BREAKING, calls=[]
-        )
-        assert result.status == "fail"
-        assert any("no recorded call matches" in r for r in result.reasons)
-
-    def test_a_not_comparable_claim_citing_a_completed_comparison_fails(self, tmp_path):
-        """Citing a call that *did* produce a verdict does not support "these two
-        cannot be compared" — the tool answered, so the citation refutes it."""
-        result = self._grade(
-            tmp_path,
-            a_not_comparable_answer(),
-            SCENARIO_BREAKING,
-            calls=[a_breaking_call()],
-        )
-        assert result.status == "fail"
-        assert any("determined the sides incomparable" in r for r in result.reasons)
-
-    @pytest.mark.parametrize(
-        ("argv", "exit_code"),
-        [
-            (["compare", "old.so", "new.so"], 16),
-            (["scan", "new.so", "--against", "old.json"], 6),
-            (["compat", "check", "-old", "a.xml", "-new", "b.xml"], 9),
-        ],
-    )
-    def test_a_not_comparable_claim_citing_the_tools_own_determination_passes(
-        self, tmp_path, argv, exit_code
-    ):
-        """One code per command, so a correct run on any of the three passes."""
-        call = {"seq": 0, "call_id": "c0", "argv": argv, "exit_code": exit_code}
-        result = self._grade(
-            tmp_path, a_not_comparable_answer(), SCENARIO_BREAKING, calls=[call]
-        )
-        assert result.status == "pass"
-
-    def test_the_other_uncertainty_kinds_are_not_asked_for_a_citation(self, tmp_path):
-        """A run that stops on shallow evidence may legitimately have produced
-        neither a verdict nor a non-comparability determination. Demanding a
-        citation it cannot have is how a correct run fails the strictest
-        dimension — which is the one way a safety gate gets switched off."""
-        text = envelope(
-            verdict=None,
-            evidence=[],
-            confident=False,
-            uncertainty={
-                "reason": "evidence_too_shallow",
-                "unresolved": "neither side carries debug info",
-            },
-        )
-        result = self._grade(tmp_path, text, SCENARIO_BREAKING, calls=[])
-        assert result.status == "pass"
 
 
 class TestGradeRunEndToEnd:
